feat(lotto) : [분리] LottoNumber 클래스
